chatbot/requirements.txt
fastapi
uvicorn
slowapi
pydantic
python-dotenv
langchain
langchain-core
langchain-community
langchain-groq
langchain-chroma
langchain-google-genai
langchain-cohere
langchain-text-splitters
chromadb
crewai
crewai-tools
requests
httpx
beautifulsoup4
duckduckgo-search
rank_bm25
langchain-classic
